v2-production/requirements.txt
boto3>=1.28.0
langchain-core==0.1.52
langsmith==0.1.17
langchain-text-splitters==0.0.1
langchain-community==0.0.38
langchain==0.1.20
langchain-aws==0.1.6
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
jupyter>=1.0.0
tqdm==4.67.1
isort==6.1.0
pre-commit==4.3.0
pymupdf==1.26.5
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv
pyyaml
Events==0.5
opensearch-py==3.0.0
requests-aws4auth==1.3.1
watchtower==3.4.0